Extract PrestoAction interface
