+++ presto-verifier/src/main/java/com/facebook/presto/verifier/framework/AbstractVerification.java
@@ -57,7 +57,7 @@ public abstract class AbstractVerification
 {
     private static final Logger log = Logger.get(AbstractVerification.class);
 
-    private final JdbcPrestoAction prestoAction;
+    private final PrestoAction prestoAction;
     private final SourceQuery sourceQuery;
     private final QueryRewriter queryRewriter;
     private final List<FailureResolver> failureResolvers;
@@ -69,7 +69,7 @@ public abstract class AbstractVerification
     private final VerificationContext verificationContext = new VerificationContext();
 
     public AbstractVerification(
-            JdbcPrestoAction prestoAction,
+            PrestoAction prestoAction,
             SourceQuery sourceQuery,
             QueryRewriter queryRewriter,
             List<FailureResolver> failureResolvers,
@@ -153,7 +153,7 @@ public abstract class AbstractVerification
         }
     }
 
-    protected JdbcPrestoAction getPrestoAction()
+    protected PrestoAction getPrestoAction()
     {
         return prestoAction;
     }
